Moved to version 1.5.0

+++ src/com/cesarandres/ps2link/module/SQLiteManager.java
@@ -64,9 +64,10 @@ public class SQLiteManager extends SQLiteOpenHelper {
 	public static final String TWEETS_COLUMN_USER = "user";
 	public static final String TWEETS_COLUMN_TAG = "tag";
 	public static final String TWEETS_COLUMN_CONTENT = "content";
+	public static final String TWEETS_COLUMN_PICTURE = "picture";
 	
 	public static final String DATABASE_NAME = "ps2link.db";
-	public static final int DATABASE_VERSION = 25;
+	public static final int DATABASE_VERSION = 28;
 
 	private static final String CREATE_WORLDS_TABLE = "create table "
 			+ TABLE_WORLDS_NAME + " ( " + WORLDS_COLUMN_ID + " Int, "
@@ -130,6 +131,7 @@ public class SQLiteManager extends SQLiteOpenHelper {
 			+ TWEETS_COLUMN_USER+ " varchar(-1), " 
 			+ TWEETS_COLUMN_TAG+ " varchar(-1), " 
 			+ TWEETS_COLUMN_CONTENT + " varchar(-1), " 
+			+ TWEETS_COLUMN_PICTURE + " varchar(-1), " 
 			+ "PRIMARY KEY (" + TWEETS_COLUMN_ID + "));";
 	
 	/**
@@ -166,12 +168,16 @@ public class SQLiteManager extends SQLiteOpenHelper {
 	 */
 	@Override
 	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_CHARACTERS_NAME);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_MEMBERS_NAME);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_OUTFITS_NAME);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_WORLDS_NAME);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_FACTIONS_NAME);
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_TWEETS_NAME);
-		onCreate(db);
+		//db.execSQL("DROP TABLE IF EXISTS " + TABLE_CHARACTERS_NAME);
+		//db.execSQL("DROP TABLE IF EXISTS " + TABLE_MEMBERS_NAME);
+		//db.execSQL("DROP TABLE IF EXISTS " + TABLE_OUTFITS_NAME);
+		//db.execSQL("DROP TABLE IF EXISTS " + TABLE_WORLDS_NAME);
+		//db.execSQL("DROP TABLE IF EXISTS " + TABLE_FACTIONS_NAME);
+		//db.execSQL("DROP TABLE IF EXISTS " + TABLE_TWEETS_NAME);
+		if(oldVersion < 28){
+			db.execSQL("DROP TABLE IF EXISTS " + TABLE_TWEETS_NAME);			
+		}
+		db.execSQL(CREATE_TWEETS_TABLE);
+		//onCreate(db);
 	}
 }
